FitParameterSet moved from Fit library to Core
