+++ Core/Fitting/IFitParameter.cpp
