+++ Core/Fitting/IFitParameter.h
