+++ Wrap/swig/libBornAgainCore.i
@@ -90,11 +90,12 @@
 #include "FTDistributions2D.h"
 #include "FitObject.h"
 #include "FitOptions.h"
+#include "IFitParameter.h"
+#include "FitParameterSet.h"
 #include "FitParameter.h"
 #include "FitSuite.h"
 #include "FitSuiteImpl.h"
 #include "FitSuiteObjects.h"
-#include "FitParameterSet.h"
 #include "FixedBinAxis.h"
 #include "FootprintFactorGaussian.h"
 #include "FootprintFactorSquare.h"
@@ -244,7 +245,6 @@
 %import(module="libBornAgainFit") "AttLimits.h"
 %import(module="libBornAgainFit") "Attributes.h"
 %import(module="libBornAgainFit") "RealLimits.h"
-%import(module="libBornAgainFit") "IFitParameter.h"
 
 %include "BAVersion.h"
 %include "BasicVector3D.h"
@@ -292,6 +292,8 @@
 %include "ChiSquaredModule.h"
 %include "FitObject.h"
 %include "FitOptions.h"
+%include "IFitParameter.h"
+%include "FitParameterSet.h"
 %include "FitParameter.h"
 %include "FitSuite.h"
 %include "FitSuiteObjects.h"
