+++ auto/Wrap/libBornAgainFit.py
@@ -2338,503 +2338,6 @@ class IMinimizer(_object):
 IMinimizer_swigregister = _libBornAgainFit.IMinimizer_swigregister
 IMinimizer_swigregister(IMinimizer)
 
-class IFitParameter(_object):
-    """
-
-
-    A fittable parameter with value, error, step, and limits.
-
-    C++ includes: IFitParameter.h
-
-    """
-
-    __swig_setmethods__ = {}
-    __setattr__ = lambda self, name, value: _swig_setattr(self, IFitParameter, name, value)
-    __swig_getmethods__ = {}
-    __getattr__ = lambda self, name: _swig_getattr(self, IFitParameter, name)
-    __repr__ = _swig_repr
-
-    def __init__(self, *args):
-        """
-        __init__(IFitParameter self) -> IFitParameter
-        __init__(IFitParameter self, std::string const & name, double value, AttLimits limits, double step=0.0) -> IFitParameter
-        __init__(IFitParameter self, std::string const & name, double value, AttLimits limits) -> IFitParameter
-        __init__(IFitParameter self, std::string const & name, double value) -> IFitParameter
-
-        IFitParameter::IFitParameter(const std::string &name, double value, const AttLimits &limits=AttLimits::limitless(), double step=0.0)
-
-        """
-        this = _libBornAgainFit.new_IFitParameter(*args)
-        try:
-            self.this.append(this)
-        except __builtin__.Exception:
-            self.this = this
-    __swig_destroy__ = _libBornAgainFit.delete_IFitParameter
-    __del__ = lambda self: None
-
-    def clone(self):
-        """
-        clone(IFitParameter self) -> IFitParameter
-
-        IFitParameter * IFitParameter::clone() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_clone(self)
-
-
-    def name(self):
-        """
-        name(IFitParameter self) -> std::string
-
-        std::string IFitParameter::name() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_name(self)
-
-
-    def setName(self, name):
-        """
-        setName(IFitParameter self, std::string const & name) -> IFitParameter
-
-        IFitParameter & IFitParameter::setName(const std::string &name)
-
-        """
-        return _libBornAgainFit.IFitParameter_setName(self, name)
-
-
-    def startValue(self):
-        """
-        startValue(IFitParameter self) -> double
-
-        double IFitParameter::startValue() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_startValue(self)
-
-
-    def setStartValue(self, value):
-        """
-        setStartValue(IFitParameter self, double value)
-
-        void IFitParameter::setStartValue(double value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setStartValue(self, value)
-
-
-    def value(self):
-        """
-        value(IFitParameter self) -> double
-
-        double IFitParameter::value() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_value(self)
-
-
-    def setValue(self, value):
-        """
-        setValue(IFitParameter self, double value)
-
-        void IFitParameter::setValue(double value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setValue(self, value)
-
-
-    def addPattern(self, pattern):
-        """
-        addPattern(IFitParameter self, std::string const & pattern) -> IFitParameter
-
-        IFitParameter & IFitParameter::addPattern(const std::string &pattern)
-
-        """
-        return _libBornAgainFit.IFitParameter_addPattern(self, pattern)
-
-
-    def step(self):
-        """
-        step(IFitParameter self) -> double
-
-        double IFitParameter::step() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_step(self)
-
-
-    def setStep(self, value):
-        """
-        setStep(IFitParameter self, double value) -> IFitParameter
-
-        IFitParameter & IFitParameter::setStep(double value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setStep(self, value)
-
-
-    def error(self):
-        """
-        error(IFitParameter self) -> double
-
-        double IFitParameter::error() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_error(self)
-
-
-    def setError(self, value):
-        """
-        setError(IFitParameter self, double value)
-
-        void IFitParameter::setError(double value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setError(self, value)
-
-
-    def limits(self, *args):
-        """
-        limits(IFitParameter self) -> AttLimits
-        limits(IFitParameter self) -> AttLimits
-
-        AttLimits & IFitParameter::limits()
-
-        """
-        return _libBornAgainFit.IFitParameter_limits(self, *args)
-
-
-    def setLimits(self, limits):
-        """
-        setLimits(IFitParameter self, AttLimits limits) -> IFitParameter
-
-        IFitParameter & IFitParameter::setLimits(const AttLimits &limits)
-
-        """
-        return _libBornAgainFit.IFitParameter_setLimits(self, limits)
-
-
-    def setLowerLimited(self, bound_value):
-        """
-        setLowerLimited(IFitParameter self, double bound_value) -> IFitParameter
-
-        IFitParameter & IFitParameter::setLowerLimited(double bound_value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setLowerLimited(self, bound_value)
-
-
-    def setPositive(self):
-        """
-        setPositive(IFitParameter self) -> IFitParameter
-
-        IFitParameter & IFitParameter::setPositive()
-
-        """
-        return _libBornAgainFit.IFitParameter_setPositive(self)
-
-
-    def setNonnegative(self):
-        """
-        setNonnegative(IFitParameter self) -> IFitParameter
-
-        IFitParameter & IFitParameter::setNonnegative()
-
-        """
-        return _libBornAgainFit.IFitParameter_setNonnegative(self)
-
-
-    def setUpperLimited(self, bound_value):
-        """
-        setUpperLimited(IFitParameter self, double bound_value) -> IFitParameter
-
-        IFitParameter & IFitParameter::setUpperLimited(double bound_value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setUpperLimited(self, bound_value)
-
-
-    def setLimited(self, left_bound_value, right_bound_value):
-        """
-        setLimited(IFitParameter self, double left_bound_value, double right_bound_value) -> IFitParameter
-
-        IFitParameter & IFitParameter::setLimited(double left_bound_value, double right_bound_value)
-
-        """
-        return _libBornAgainFit.IFitParameter_setLimited(self, left_bound_value, right_bound_value)
-
-
-    def setFixed(self):
-        """
-        setFixed(IFitParameter self) -> IFitParameter
-
-        IFitParameter & IFitParameter::setFixed()
-
-        """
-        return _libBornAgainFit.IFitParameter_setFixed(self)
-
-
-    def toString(self):
-        """
-        toString(IFitParameter self) -> std::string
-
-        std::string IFitParameter::toString() const 
-
-        """
-        return _libBornAgainFit.IFitParameter_toString(self)
-
-IFitParameter_swigregister = _libBornAgainFit.IFitParameter_swigregister
-IFitParameter_swigregister(IFitParameter)
-
-class FitParameterSet(_object):
-    """
-
-
-    The set of fit parameters.
-
-    C++ includes: FitParameterSet.h
-
-    """
-
-    __swig_setmethods__ = {}
-    __setattr__ = lambda self, name, value: _swig_setattr(self, FitParameterSet, name, value)
-    __swig_getmethods__ = {}
-    __getattr__ = lambda self, name: _swig_getattr(self, FitParameterSet, name)
-    __repr__ = _swig_repr
-
-    def __init__(self):
-        """
-        __init__(FitParameterSet self) -> FitParameterSet
-
-        FitParameterSet::FitParameterSet()
-
-        """
-        this = _libBornAgainFit.new_FitParameterSet()
-        try:
-            self.this.append(this)
-        except __builtin__.Exception:
-            self.this = this
-    __swig_destroy__ = _libBornAgainFit.delete_FitParameterSet
-    __del__ = lambda self: None
-
-    def clear(self):
-        """
-        clear(FitParameterSet self)
-
-        void FitParameterSet::clear()
-
-        Clears all defined parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_clear(self)
-
-
-    def size(self):
-        """
-        size(FitParameterSet self) -> size_t
-
-        size_t FitParameterSet::size() const
-
-        Returns number of parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_size(self)
-
-
-    def begin(self, *args):
-        """
-        begin(FitParameterSet self) -> FitParameterSet::iterator
-        begin(FitParameterSet self) -> FitParameterSet::const_iterator
-
-        FitParameterSet::const_iterator FitParameterSet::begin() const 
-
-        """
-        return _libBornAgainFit.FitParameterSet_begin(self, *args)
-
-
-    def end(self, *args):
-        """
-        end(FitParameterSet self) -> FitParameterSet::iterator
-        end(FitParameterSet self) -> FitParameterSet::const_iterator
-
-        FitParameterSet::const_iterator FitParameterSet::end() const 
-
-        """
-        return _libBornAgainFit.FitParameterSet_end(self, *args)
-
-
-    def addFitParameter(self, par):
-        """
-        addFitParameter(FitParameterSet self, IFitParameter par)
-
-        void FitParameterSet::addFitParameter(IFitParameter *par)
-
-        Adds fit parameter. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_addFitParameter(self, par)
-
-
-    def fitParameter(self, *args):
-        """
-        fitParameter(FitParameterSet self, std::string const & name) -> IFitParameter
-        fitParameter(FitParameterSet self, std::string const & name) -> IFitParameter
-
-        IFitParameter * FitParameterSet::fitParameter(const std::string &name)
-
-        """
-        return _libBornAgainFit.FitParameterSet_fitParameter(self, *args)
-
-
-    def values(self):
-        """
-        values(FitParameterSet self) -> vdouble1d_t
-
-        std::vector< double > FitParameterSet::values() const
-
-        Returns values of all defined parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_values(self)
-
-
-    def setValues(self, pars_values):
-        """
-        setValues(FitParameterSet self, vdouble1d_t pars_values)
-
-        void FitParameterSet::setValues(const std::vector< double > &pars_values)
-
-        Sets values for all defined parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_setValues(self, pars_values)
-
-
-    def valuesDifferFrom(self, par_values, tolerance=2.0):
-        """
-        valuesDifferFrom(FitParameterSet self, vdouble1d_t par_values, double tolerance=2.0) -> bool
-        valuesDifferFrom(FitParameterSet self, vdouble1d_t par_values) -> bool
-
-        bool FitParameterSet::valuesDifferFrom(const std::vector< double > &par_values, double tolerance=2.0) const
-
-        Returns true if parameters already have the given values. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_valuesDifferFrom(self, par_values, tolerance)
-
-
-    def errors(self):
-        """
-        errors(FitParameterSet self) -> vdouble1d_t
-
-        std::vector< double > FitParameterSet::errors() const
-
-        Returns errors of all defined parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_errors(self)
-
-
-    def setErrors(self, pars_errors):
-        """
-        setErrors(FitParameterSet self, vdouble1d_t pars_errors)
-
-        void FitParameterSet::setErrors(const std::vector< double > &pars_errors)
-
-        Sets errors to all parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_setErrors(self, pars_errors)
-
-
-    def freeFitParameterCount(self):
-        """
-        freeFitParameterCount(FitParameterSet self) -> size_t
-
-        size_t FitParameterSet::freeFitParameterCount() const
-
-        Returns number of free parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_freeFitParameterCount(self)
-
-
-    def fixAll(self):
-        """
-        fixAll(FitParameterSet self)
-
-        void FitParameterSet::fixAll()
-
-        Fix all parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_fixAll(self)
-
-
-    def releaseAll(self):
-        """
-        releaseAll(FitParameterSet self)
-
-        void FitParameterSet::releaseAll()
-
-        Release all parameters. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_releaseAll(self)
-
-
-    def setFixed(self, pars, is_fixed):
-        """
-        setFixed(FitParameterSet self, vector_string_t pars, bool is_fixed)
-
-        void FitParameterSet::setFixed(const std::vector< std::string > &pars, bool is_fixed)
-
-        Set fixed flag for parameters from the list. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_setFixed(self, pars, is_fixed)
-
-
-    def correlationMatrix(self):
-        """
-        correlationMatrix(FitParameterSet self) -> vdouble2d_t
-
-        corr_matrix_t FitParameterSet::correlationMatrix() const 
-
-        """
-        return _libBornAgainFit.FitParameterSet_correlationMatrix(self)
-
-
-    def setCorrelationMatrix(self, matrix):
-        """
-        setCorrelationMatrix(FitParameterSet self, vdouble2d_t matrix)
-
-        void FitParameterSet::setCorrelationMatrix(const corr_matrix_t &matrix)
-
-        Sets resulting correlation matrix. 
-
-        """
-        return _libBornAgainFit.FitParameterSet_setCorrelationMatrix(self, matrix)
-
-
-    def fitParametersNewKernel(self):
-        """fitParametersNewKernel(FitParameterSet self) -> Parameters"""
-        return _libBornAgainFit.FitParameterSet_fitParametersNewKernel(self)
-
-
-    def __getitem__(self, *args):
-        """
-        __getitem__(FitParameterSet self, std::string name) -> IFitParameter
-        __getitem__(FitParameterSet self, size_t index) -> IFitParameter
-        """
-        return _libBornAgainFit.FitParameterSet___getitem__(self, *args)
-
-
-    def __iter__(self):
-        return FitParameterSetIterator(self)
-
-FitParameterSet_swigregister = _libBornAgainFit.FitParameterSet_swigregister
-FitParameterSet_swigregister(FitParameterSet)
-
 class MinimizerCatalogue(_object):
     """
